pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "nonebot-plugin-course-schedule"
version = "1.1.1"
description = "电子课程表"
authors = [
    {name = "Polaris_Light", email = "995905922@qq.com" },
    {name = "Jason.Joestar"}
]
dependencies = [
    "nonebot2[fastapi]>=2.2.1",
    "nonebot-adapter-onebot>=2.0.0-beta.1",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-localstore>=0.7.4",
    "pydantic>=1.10",
    "aiohttp>=3.8.0",
    "Pillow>=11.0.0",
    "icalendar>=6.3.1",
    "python-dateutil>=2.8.0",
    "jionlp>=1.5.27",
]
readme = "README.md"
license = "AGPL-3.0"
license-files = ["LICEN[CS]E*"]
requires-python = ">=3.10"
keywords = ["课程", "course"]
classifiers = []

[project.urls]
Homepage = "https://github.com/GLDYM/nonebot-plugin-course-schedule"

[tool.poetry.dependencies]
nonebot2 = ">=2.2.1"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
nonebot-plugin-apscheduler = ">=0.4.0"
nonebot-plugin-localstore = ">=0.7.4"
pydantic = ">=1.10"
aiohttp = ">=3.8.0"
Pillow = ">=11.0.0"
icalendar = ">=6.3.1"
python-dateutil = ">=2.8.0"
jionlp = ">=1.5.27"

[tool.nonebot.plugins]
nonebot-plugin-course-schedule = ["nonebot_plugin_course_schedule"]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
